requirements.txt

Django==4.2.18
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==23.0.0
